+++ CHANGES.rst
@@ -5,8 +5,11 @@ Changes
 -------------------
 * Support for multiple module colors
 * Support for Segno's API >= 1.0.0
-* Changed parameter "color" to "dark" and "background" to "light"
-  to match Segno's API
+* API breaking changes:
+
+  - Changed parameter "color" to "dark" and "background" to "light"
+    to match Segno's API
+  - Removed "mode" parameter
 
 
 0.1.6 -- 2016-09-19